Create Powershell-Case Folder Structure Creation
a script to create the case folder structure I prefer. Main folder is the case number - the suspects surname. Within that main folder are 3 subfolders. Images subfolder to store all the working forensic images organized by a sub folder for each evidence item. Media subfolder with collects the VICs media exports from each evidence item to later be processed by Griffeye app, and lastly a Griffeye folder to hold the Griffeye case data. Griffeye is the only tool that is about universal for me so each of the other tools files I'll create when i call up the tool in their corresponding scripts.